requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
PyYAML==6.0.2
python-dateutil==2.9.0.post0
requests==2.32.3
